+++ alexis-vilai-photography/public/manifest.json
