+++ alexis-vilai-photography/src/serviceWorker.js
